perf(triattn-bwd): store dq grad bf16 -> kills last cast copy (v7, 1.78x)
dq was fp32 as a vestige of the original atomic-add dq (atomics needed an fp32
accumulator buffer). The v2 FA2 split removed the atomics but kept the fp32 dtype
by inheritance. _attn_bwd_dq accumulates in fp32 registers and stores once like
dk/dv, so bf16 out is precision-neutral (ref stores bf16 grads too). Kills the
backward copy_ [1,4,384,384,128] fp32->bf16 cast. cos 0.99999/0.99998 unchanged;
step 6.403->6.165ms. All wasteful copies now eliminated (only intentional
bias.contiguous remains).

+++ src/miniworld_kernels/kernels/triangle_attention/triton/main.py
@@ -382,7 +382,12 @@ class TritonTriangleAttentionPairBiasFunction(torch.autograd.Function):
         # A: allocate dq/dk/dv in the PROJECTION layout [B,L,L2,H*D] and hand the kernels
         # strided (B,H,L,L2,D) views of them. The kernels write strided (q-group), so the
         # module's rearrange-backward becomes a FREE view -> no grad-transpose copy.
-        dq_proj = torch.empty(B, L, L, H * D, device=q.device, dtype=torch.float32)
+        # dq stored bf16 (was fp32 — a vestige of the original atomic-add dq, which needed an
+        # fp32 accumulator buffer; FA2 split removed the atomics). The kernel accumulates dq in
+        # fp32 registers and stores once, exactly like dk/dv, so bf16 out is precision-neutral
+        # (matches the cuBLAS/pytorch ref which also stores bf16 grads). Kills the fp32->bf16
+        # cast copy of the returned query grad.
+        dq_proj = torch.empty(B, L, L, H * D, device=q.device, dtype=v.dtype)
         dk_proj = torch.empty(B, L, L, H * D, device=q.device, dtype=v.dtype)
         dv_proj = torch.empty(B, L, L, H * D, device=q.device, dtype=v.dtype)
         dq = rearrange(dq_proj, "B L L2 (H D) -> B H L L2 D", H=H)
